[LayoutNG] OOF fragmentainer descendant conversions
If there was no containing block fragment set on an OOF fragmentainer
descendant, we were using an empty size for conversions in
NGPhysicalBoxFragment. This is incorrect - when no containing block is
set, that means the current fragment is the containing block. Use the
size of the current fragment instead.

Additionally, an incorrect converter was used in converting the static
position of OOF fragmentainer descendants in NGPhysicalBoxFragment.
This CL ensures that the static position remains relative to the
containing block, similar to how the conversions are handled in
NGContainerFragmentBuilder.

Bug: 1227115
Change-Id: Idb455573b1bf6b23c80679d8494b56ae267cf34a